Connect: re-align root CA state if needed
This prevents state to drift if structs.CARoot gets modified and a new
field is added.
It is important to ensure results from /v1/connect/ca/roots, which
return data from state, are consistent. For instance, on a cluster
that has been setup in 2018, we observed empty ExternalTrustDomain,
as well as PrivateKeyType and PrivateKeyBits.
Without this patch, one would need to wait a rotation to fix state,
which happens every 5 years by default.
Also address TODO that was here now we have strong guarantees that
rootCA and activeRoot are the same, and add a test to ensure
idempotency.